[FIX] 잠금 모드 설정 중 취소한 경우
