+++ app/src/main/java/com/example/main_dialendar/view/fragment/SettingPreferenceFragment.java
@@ -12,10 +12,13 @@ import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
 import android.util.Log;
-import android.widget.BaseAdapter;
 import android.widget.TimePicker;
 import android.widget.Toast;
 
+import androidx.activity.result.ActivityResult;
+import androidx.activity.result.ActivityResultCallback;
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.preference.ListPreference;
 import androidx.preference.Preference;
 import androidx.preference.PreferenceFragmentCompat;
@@ -26,6 +29,7 @@ import com.example.main_dialendar.R;
 import com.example.main_dialendar.util.message.MessageReceiver;
 import com.example.main_dialendar.util.setting.SharedPrefManager;
 import com.example.main_dialendar.util.theme.ThemeUtil;
+import com.example.main_dialendar.view.activity.LockActivity;
 import com.example.main_dialendar.view.activity.SettingActivity;
 import com.example.main_dialendar.view.dialog.TimePickerDialog;
 
@@ -60,8 +64,12 @@ public class SettingPreferenceFragment extends PreferenceFragmentCompat {
         setPreferencesFromResource(R.xml.settings_preference, rootKey);  // xml 파일과 연동
         context = getActivity();
 
-        localPref = PreferenceManager.getDefaultSharedPreferences(context);
+        setLocalPreferences();
+        connectWithSharedPreferences();
+    }
 
+    private void setLocalPreferences() {
+        localPref = PreferenceManager.getDefaultSharedPreferences(context);
         messagePref = findPreference("message");
         messagePref.setOnPreferenceChangeListener(onPreferenceChangeListener);
         lockPref = findPreference("lock");
@@ -69,14 +77,14 @@ public class SettingPreferenceFragment extends PreferenceFragmentCompat {
         darkmodePref = findPreference("darkmode");
         darkmodePref.setOnPreferenceChangeListener(onPreferenceChangeListener);
 
-        // sharedPreferences 연결
+        // 디폴트 값 설정
+        setDefaultInPrefs();
+    }
+
+    private void connectWithSharedPreferences() {
         sharedPref = SharedPrefManager.getInstance(context);
         messageHour = sharedPref.getMessageHour();
         messageMinute = sharedPref.getMessageMinute();
-
-        // 디폴트 값 설정
-        setDefaultInPrefs();
-
     }
 
     private void setDefaultInPrefs() {
@@ -133,10 +141,49 @@ public class SettingPreferenceFragment extends PreferenceFragmentCompat {
     private void setLockmode() {
         SettingActivity settingActivity = (SettingActivity) getActivity();
         if (localPref.getBoolean("lock", false))
-            settingActivity.moveToLockActivity(LOCKMODE_ON);
+            moveToLockActivity(LOCKMODE_ON);
         else
-            settingActivity.moveToLockActivity(LOCKMODE_OFF);
+            moveToLockActivity(LOCKMODE_OFF);
     }
+
+    public void moveToLockActivity(int mode) {
+        Intent intent = new Intent(context, LockActivity.class);
+        intent.putExtra("lock", mode);
+        lockOnOffLauncher.launch(intent);
+    }
+
+    ActivityResultLauncher<Intent> lockOnOffLauncher = registerForActivityResult(
+                new ActivityResultContracts.StartActivityForResult(),
+                new ActivityResultCallback<ActivityResult>() {
+        @Override
+        public void onActivityResult(ActivityResult result) {
+            Log.e("###", "Fragment에서 실행");
+
+            int resultCode = result.getResultCode();
+            int mode = result.getData().getIntExtra("mode", LOCKMODE_OFF);
+
+            switch(mode) {
+                case LOCKMODE_ON:
+                    if (resultCode == RESULT_OK) {
+                        lockPref.setSummary("사용");
+                        sharedPref.setLockOn(true);
+                    }
+                    else
+                        lockPref.setChecked(false);
+                    break;
+                case LOCKMODE_OFF:
+                    if (resultCode == RESULT_OK) {
+                        lockPref.setSummary("사용 안 함");
+                        sharedPref.setLockOn(false);
+                    }
+                    else
+                        lockPref.setChecked(true);
+                    break;
+            }
+        }
+    });
+
+
 
     private void showMessageDialog() {
         TimePickerDialog timePickerDialog = new TimePickerDialog(context);
@@ -145,31 +192,6 @@ public class SettingPreferenceFragment extends PreferenceFragmentCompat {
         timePickerDialog.setCanceledOnTouchOutside(false);
         timePickerDialog.setCancelable(true);
         timePickerDialog.show();
-    }
-
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        Log.e("###", "Fragment: onActivityResult!!");
-
-        switch(data.getIntExtra("mode", LOCKMODE_OFF)) {
-            case LOCKMODE_ON:
-                if (resultCode == RESULT_OK) {
-                    lockPref.setSummary("사용");
-                    sharedPref.setLockOn(true);
-                }
-                else
-                    lockPref.setChecked(false);
-                break;
-            case LOCKMODE_OFF:
-                if (resultCode == RESULT_OK) {
-                    lockPref.setSummary("사용 안 함");
-                    sharedPref.setLockOn(false);
-                }
-                else
-                    lockPref.setChecked(true);
-                break;
-        }
     }
 
     android.app.TimePickerDialog.OnTimeSetListener timeSetListener = new android.app.TimePickerDialog.OnTimeSetListener() {
